persist loksabha 15 531 mps data
As part of this commit. All the MP data is persisted for loksabha 15.
Keys were modified to match the db columns in mps.json. Still
Insert_batch to be used to make writes faster. But, it is advisable not
to use insert_Batch if there are more than 10,000 records. Under data
set loksabha_15_531_mps.json is placed, this is our master json file.